#11 [feat] kakao social 로그인/회원가입

+++ server/src/main/java/org/tattour/server/global/config/jwt/JwtService.java
@@ -28,7 +28,7 @@ public class JwtService {
 	}
 
 	// JWT 토큰 발급
-	public String issuedToken(Long userId) {
+	public String issuedToken(Integer userId) {
 		String strUserId = String.valueOf(userId);
 		final Date now = new Date();
